Move network policies into their own file
